+++ opentelemetry-api/src/opentelemetry/util/__init__.py
@@ -19,6 +19,8 @@ from typing import TYPE_CHECKING, Union, cast
 
 from pkg_resources import iter_entry_points
 
+from opentelemetry.environment_variables import OTEL_PYTHON_TRACER_PROVIDER
+
 if TYPE_CHECKING:
     from opentelemetry.trace import TracerProvider
 
@@ -38,7 +40,9 @@ except AttributeError:
         return int(time.time() * 1e9)
 
 
-def _load_provider(provider: str) -> Provider:
+def _load_provider(
+    provider_environment_variable: str, provider: str
+) -> Provider:
     try:
         entry_point = next(
             iter_entry_points(
@@ -46,7 +50,8 @@ def _load_provider(provider: str) -> Provider:
                 name=cast(
                     str,
                     environ.get(
-                        provider.upper(), "default_{}".format(provider),
+                        provider_environment_variable,
+                        "default_{}".format(provider),
                     ),
                 ),
             )
@@ -58,4 +63,7 @@ def _load_provider(provider: str) -> Provider:
 
 
 def _load_trace_provider(provider: str) -> "TracerProvider":
-    return cast("TracerProvider", _load_provider(provider))  # type: ignore
+    return cast(
+        "TracerProvider",
+        _load_provider(OTEL_PYTHON_TRACER_PROVIDER, provider),
+    )  # type: ignore